% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data1}
\alias{data1}
\title{One-dimensional gaussian segmentation}
\usage{
data1(mean = c(0.5, 0, -0.5, 0), tau = c(0.25, 0.5, 0.75, 1), sigma = 1,
  n = 100)
}
\arguments{
\item{mean}{Vector of segment means}

\item{tau}{relative position (between 0 and 1) of the changepoints (the last position is always 1)}

\item{sigma}{the value of the standard deviation for the simulation}

\item{n}{The number of data to generate}
}
\value{
A vector of data of length n generated by the simulated segmentation
}
\description{
Simulating a one-dimenstional segmentation with given number of segments, their location, mean and standard deviation (constant)
}
\examples{
data <- data1(sigma = 0.05, n=20)
plot(1:20,data)
}
