% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data2}
\alias{data2}
\title{Two-dimensional gaussian segmentation}
\usage{
data2(mean1 = c(0, 1, 1, 0), mean2 = c(0, 0, 1, 1), tau = c(0.25, 0.5,
  0.75, 1), sigma = 1, n = 100)
}
\arguments{
\item{mean1}{Vector of segment means for dimension 1}

\item{mean2}{Vector of segment means for dimension 2}

\item{tau}{relative position (between 0 and 1) of the changepoints (the last position is always 1)}

\item{sigma}{the value of the standard deviation for the data}

\item{n}{The number of data to generate}
}
\value{
A matrix of data of dimension 2 x n generated by the simulated segmentation
}
\description{
Simulating a two-dimenstional segmentation with given number of segments, their location, mean and standard deviation (constant)
}
\examples{
data <- data2(sigma = 0.05, n=20)
max = max(data);min = min(data)
plot(1:20,data[1,],ylim = c(min,max))
par(new = TRUE)
plot(1:20,data[2,],ylim = c(min,max),col=2)
par(new = FALSE)
}
