% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataG.R
\name{dataAll}
\alias{dataAll}
\title{data in dimension 1 with no pruned last changepoint candidate}
\usage{
dataAll(n, beta = 1)
}
\arguments{
\item{n}{length of the vector}

\item{beta}{the penalty coefficient (use the same beta in function fpop1d)}
}
\value{
A vector of data of length n
}
\description{
Generating data in dimension 1 with no pruning
}
\examples{
data <- dataAll(n = 10, beta = 2)
fpop1d(data, beta = 2)


}
