% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataG.R
\name{dataG1}
\alias{dataG1}
\title{One-dimensional gaussian segmentation}
\usage{
dataG1(mean = c(0.5, 0, -0.5, 0), tau = c(0.25, 0.5, 0.75, 1), sigma = 1,
  n = 100)
}
\arguments{
\item{mean}{Vector of means (mean of consecutive segments)}

\item{tau}{relative position (between 0 and 1) of the changepoints (the last position is always 1)}

\item{sigma}{A positive number: the value (unique) of the standard deviation for all the Gaussian laws in the simulation}

\item{n}{The number of data point to simulate}
}
\value{
A vector of data of length n generated by the simulated segmentation
}
\description{
Simulating a one-dimenstional segmentation with a given number of segments
}
\examples{
data <- dataG1(sigma = 0.05, n=20)
plot(1:20,data)
}
