% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataG.R
\name{dataG2}
\alias{dataG2}
\title{Two-dimensional gaussian segmentation}
\usage{
dataG2(mean1 = c(0, 1, 1, 0), mean2 = c(0, 0, 1, 1), tau = c(0.25, 0.5,
  0.75, 1), sigma = 1, n = 100)
}
\arguments{
\item{mean1}{Vector of means for the first dimension}

\item{mean2}{Vector of means for the second dimension}

\item{tau}{relative position (between 0 and 1) of the changepoints (the last position is always 1)}

\item{sigma}{A positive number: the value (unique) of the standard deviation for all the Gaussian laws in the simulation #' @param n The number of data point to simulate}

\item{n}{The number of data point to simulate}
}
\value{
A matrix of data of dimension 2 x n generated by the simulated segmentation
}
\description{
Simulating a two-dimenstional segmentation with a given number of segments
}
\examples{
data <- dataG2(sigma = 0.05, n=50)
max = max(data);min = min(data)
plot(1:50,data[1,],ylim = c(min,max))
par(new = TRUE)
plot(1:50,data[2,],ylim = c(min,max),col=2)
par(new = FALSE)
}
