% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpop1d.R
\name{fpop1d}
\alias{fpop1d}
\title{Functional cost 1d display}
\usage{
fpop1d(data1, beta)
}
\arguments{
\item{data1}{Vector of data in dim 1}

\item{beta}{penalty positive number}
}
\value{
Display the cost (in the plot window) and the present labels (in the console) at consecutive times
}
\description{
Displaying the functional cost at consecutive time
}
\examples{
data <- data1(sigma = 1, n=20)
fpop1d(data,1)
}
