% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpop2d.R
\name{fpop2d}
\alias{fpop2d}
\title{Functional cost 2d display at level beta}
\usage{
fpop2d(data2, beta, precision = 300, circle = FALSE)
}
\arguments{
\item{data2}{matrix of data of dimension 2 x n}

\item{beta}{penalty coefficient, positive number}

\item{precision}{An integer : the number of rows in the matrix for plotting}

\item{circle}{An boolean to decide to draw the circle of intersection (green if the region stays, red otherwise)}
}
\value{
Display the cost (in the plot window) and the present labels (in the console) at consecutive times
}
\description{
Displaying the functional cost at consecutive times
}
\examples{
data <- data2(mean1 = 0, mean2 = 0, tau = 1, sigma = 0.5, n=10)
fpop2d(data, precision = 300, 1, circle = TRUE)


}
