% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpop1d.R
\name{fpop1d}
\alias{fpop1d}
\title{Functional cost 1d display}
\usage{
fpop1d(data1, beta, order = FALSE)
}
\arguments{
\item{data1}{Vector of data}

\item{beta}{penalty coefficient, positive number}

\item{order}{a boolean. If true, gives the labels on the real line from left to right}
}
\value{
plots of the functional cost (in the plot window) and the present labels (in the console) at consecutive times. The vertical gray line shows the position of the current added data on the real line
}
\description{
Displaying the functional cost at consecutive times
}
\examples{
data <- dataG1(sigma = 0.5, n=15)
fpop1d(data,1)
}
